Add RmsNormQunat Runner

+++ scripts/build.sh
@@ -29,10 +29,14 @@ DEVICE_CODE_PACK_SWITCH=ON
 USE_CXX11_ABI=ON
 USE_VERBOSE=OFF
 BUILD_OPTION_LIST="3rdparty download_testdata unittest unittest_and_run pythontest pythontest_and_run debug release help examples"
-BUILD_CONFIGURE_LIST=("--output=.*" "--cache=.*" "--verbose" "--incremental" "--gcov" "--no_hostbin" "--no_devicebin" "--use_cxx11_abi=0" "--use_cxx11_abi=1" "--build_config=.*" "--optimize_off" "--use_torch_runner" "--use_lccl_runner" "--use_profiling")
+BUILD_CONFIGURE_LIST=("--output=.*" "--cache=.*" "--verbose" "--incremental" "--gcov" "--no_hostbin" "--no_devicebin" "--use_cxx11_abi=0" 
+    "--use_cxx11_abi=1" "--build_config=.*" "--optimize_off" "--use_torch_runner" "--use_lccl_runner" "--use_hccl_runner" "--use_profiling")
 
 function fn_build_googltest()
 {
+    if [ -d "$THIRD_PARTY_DIR/googletest/lib" -a -d "$THIRD_PARTY_DIR/googletest/include" ];then
+        return $?
+    fi
     cd $CACHE_DIR
     rm -rf v1.13.0.tar.gz
     if [ -f "$CODE_ROOT/3rdparty/googletest-1.13.0.tar.gz" ];then
@@ -41,19 +45,31 @@ function fn_build_googltest()
         wget --no-check-certificate https://github.com/google/googletest/archive/refs/tags/v1.13.0.tar.gz
     fi
 
-    tar -xvf v1.13.0.tar.gz
+    tar -xf v1.13.0.tar.gz
     cd googletest-1.13.0
     mkdir build
     cd build
-    cmake -DCMAKE_INSTALL_PREFIX=$THIRD_PARTY_DIR/googletest ../
+    if [ "$USE_CXX11_ABI" == "ON" ]
+    then
+        sed -i '4 a add_compile_definitions(_GLIBCXX_USE_CXX11_ABI=1)' ../CMakeLists.txt
+    else
+        sed -i '4 a add_compile_definitions(_GLIBCXX_USE_CXX11_ABI=0)' ../CMakeLists.txt
+    fi
+    cmake -DCMAKE_INSTALL_PREFIX=$THIRD_PARTY_DIR/googletest -DBUILD_SHARED_LIBS=ON ../
     make -j
     make install
     cd ../../
+    if [ -d "$THIRD_PARTY_DIR/googletest/lib64" ];then
+        mv $THIRD_PARTY_DIR/googletest/lib64 $THIRD_PARTY_DIR/googletest/lib
+    fi
 }
-
 
 function fn_build_half()
 {
+    if [ -d "$THIRD_PARTY_DIR/half/include" ];then
+        echo "half exist in $THIRD_PARTY_DIR, not build it"
+        return
+    fi
     cd $CACHE_DIR
     rm -rf half-2.2.0.zip
     if [ -f "$CODE_ROOT/3rdparty/half-2.2.0.zip" ];then
@@ -73,6 +89,8 @@ function fn_build_asdops()
         echo "asdops exist in $THIRD_PARTY_DIR, not build it"
         return
     fi
+
+    cd $CACHE_DIR
     git clone https://gitee.com/ascend/ascend-op-common-lib.git
     cd ascend-op-common-lib
     ASD_OPP_PATH=/usr/local/Ascend/ascend-toolkit/6.0.2/opp
@@ -89,6 +107,8 @@ function fn_build_asdops()
     build_options="$build_options --output=$THIRD_PARTY_DIR"
     echo "bash scripts/build.sh dev $build_options"
     bash scripts/build.sh dev $build_options
+    cd $CACHE_DIR
+    sudo rm -rf ascend-op-common-lib
 }
 
 function fn_build_nlohmann_json()
@@ -103,6 +123,8 @@ function fn_build_nlohmann_json()
         unzip include.zip
         mkdir -p $THIRD_PARTY_DIR/nlohmannJson
         cp -r ./include $THIRD_PARTY_DIR/nlohmannJson
+        cd $CACHE_DIR
+        rm -rf nlohmann
     fi
 }
 
@@ -112,6 +134,7 @@ function fn_build_3rdparty()
     mkdir $CACHE_DIR
     cd $CACHE_DIR
     fn_build_googltest
+    fn_build_half
     fn_build_nlohmann_json
     fn_build_asdops
     cd ..
@@ -253,7 +276,11 @@ function fn_generate_doxygen()
     if [[ $OUTPUT_DIR != $CODE_ROOT/output ]];then
         sed -i 's|OUTPUT_DIRECTORY       =.\/output\/acltransformer\/doc|OUTPUT_DIRECTORY       ='"$OUTPUT_DIR"'\/acltransformer\/doc|g' $CODE_ROOT/Doxyfile
     fi
-    /usr/local/doxygen/bin/doxygen $CODE_ROOT/Doxyfile >/dev/null 2>&1
+    if [ -f "/usr/local/doxygen/bin/doxygen" ];then
+        /usr/local/doxygen/bin/doxygen $CODE_ROOT/Doxyfile
+    else
+        echo "/usr/local/doxygen/bin/doxygen not exist, not generate doc"
+    fi
 }
 
 function fn_run_pythontest()
@@ -388,6 +415,9 @@ function fn_main()
         "--use_lccl_runner")
             COMPILE_OPTIONS="${COMPILE_OPTIONS} -DUSE_LCCL_RUNNER=ON"
             ;;
+        "--use_hccl_runner")
+            COMPILE_OPTIONS="${COMPILE_OPTIONS} -DUSE_HCCL_RUNNER=ON"
+            ;;
         "--use_profiling")
             COMPILE_OPTIONS="${COMPILE_OPTIONS} -DUSE_PROFILING=ON"
             ;;
@@ -398,9 +428,6 @@ function fn_main()
 
     fn_init_pytorch_env
     case "${arg1}" in
-        "3rdparty")
-            fn_build_3rdparty
-            ;;
         "download_testdata")
             fn_download_testdata
             ;;
@@ -410,12 +437,15 @@ function fn_main()
             ;;
         "unittest")
             COMPILE_OPTIONS="${COMPILE_OPTIONS} -DUSE_UT_TEST=ON"
+            fn_build_3rdparty
             fn_build
             ;;
         "unittest_and_run")
             COMPILE_OPTIONS="${COMPILE_OPTIONS} -DUSE_UT_TEST=ON"
             COMPILE_OPTIONS="${COMPILE_OPTIONS} -DUSE_GCOV=ON"
+            fn_build_3rdparty
             fn_build
+            fn_run_unittest
             fn_build_coverage
             ;;
         "pythontest")
